feat: refactor tests into ProcessInstanceAndFlowNodeInstanceQueryTest
- reduce test execution time
- add utility class for test queries
