+++ qa/integration-tests/src/test/java/io/camunda/it/client/QueryTest.java
@@ -1,0 +1,128 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.it.client;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import io.camunda.zeebe.client.ZeebeClient;
+import io.camunda.zeebe.client.api.command.CreateProcessInstanceCommandStep1;
+import io.camunda.zeebe.client.api.response.DeploymentEvent;
+import io.camunda.zeebe.client.api.response.ProcessInstanceEvent;
+import io.camunda.zeebe.client.api.search.response.SearchQueryResponse;
+import java.time.Duration;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Objects;
+import java.util.function.Function;
+import org.awaitility.Awaitility;
+
+abstract class QueryTest {
+
+  static DeploymentEvent deployResource(final ZeebeClient zeebeClient, final String resourceName) {
+    return zeebeClient
+        .newDeployResourceCommand()
+        .addResourceFromClasspath(resourceName)
+        .send()
+        .join();
+  }
+
+  static ProcessInstanceEvent startProcessInstance(
+      final ZeebeClient zeebeClient, final String bpmnProcessId) {
+    return startProcessInstance(zeebeClient, bpmnProcessId, null);
+  }
+
+  static ProcessInstanceEvent startProcessInstance(
+      final ZeebeClient zeebeClient, final String bpmnProcessId, final String payload) {
+    final CreateProcessInstanceCommandStep1.CreateProcessInstanceCommandStep3
+        createProcessInstanceCommandStep3 =
+            zeebeClient.newCreateInstanceCommand().bpmnProcessId(bpmnProcessId).latestVersion();
+    if (payload != null) {
+      createProcessInstanceCommandStep3.variables(payload);
+    }
+    return createProcessInstanceCommandStep3.send().join();
+  }
+
+  static void waitForProcessInstancesToStart(
+      final ZeebeClient zeebeClient, final int expectedProcessInstances) {
+    Awaitility.await("should start process instances and import in Operate")
+        .atMost(Duration.ofSeconds(60))
+        .ignoreExceptions() // Ignore exceptions and continue retrying
+        .untilAsserted(
+            () -> {
+              final var result = zeebeClient.newProcessInstanceQuery().send().join();
+              assertThat(result.items().size()).isEqualTo(expectedProcessInstances);
+            });
+  }
+
+  static void waitForFlowNodeInstances(
+      final ZeebeClient zeebeClient, final int expectedFlowNodeInstances) {
+    Awaitility.await("should wait until flow node instances are available")
+        .atMost(Duration.ofSeconds(60))
+        .ignoreExceptions() // Ignore exceptions and continue retrying
+        .untilAsserted(
+            () -> {
+              final var result = zeebeClient.newFlownodeInstanceQuery().send().join();
+              assertThat(result.items().size()).isEqualTo(expectedFlowNodeInstances);
+            });
+  }
+
+  static void waitForProcessesToBeDeployed(
+      final ZeebeClient zeebeClient, final int expectedProcessDefinitions) {
+    Awaitility.await("should deploy processes and import in Operate")
+        .atMost(Duration.ofSeconds(15))
+        .ignoreExceptions() // Ignore exceptions and continue retrying
+        .untilAsserted(
+            () -> {
+              final var result = zeebeClient.newProcessDefinitionQuery().send().join();
+              assertThat(result.items().size()).isEqualTo(expectedProcessDefinitions);
+            });
+  }
+
+  static void waitUntilProcessInstanceHasIncidents(
+      final ZeebeClient zeebeClient, final int expectedIncidents) {
+    Awaitility.await("should wait until incidents are exists")
+        .atMost(Duration.ofSeconds(15))
+        .ignoreExceptions() // Ignore exceptions and continue retrying
+        .untilAsserted(
+            () -> {
+              final var result =
+                  zeebeClient
+                      .newProcessInstanceQuery()
+                      .filter(f -> f.hasIncident(true))
+                      .send()
+                      .join();
+              assertThat(result.items().size()).isEqualTo(expectedIncidents);
+            });
+  }
+
+  static <T, U extends Comparable<U>> void assertSorted(
+      final SearchQueryResponse<T> resultAsc,
+      final SearchQueryResponse<T> resultDesc,
+      final Function<T, U> propertyExtractor) {
+    assertThatIsAscSorted(resultAsc.items(), propertyExtractor);
+    assertThatIsDescSorted(resultDesc.items(), propertyExtractor);
+  }
+
+  static <T, U extends Comparable<U>> void assertThatIsAscSorted(
+      final List<T> items, final Function<T, U> propertyExtractor) {
+    assertThatIsSortedBy(items, propertyExtractor, Comparator.naturalOrder());
+  }
+
+  static <T, U extends Comparable<U>> void assertThatIsDescSorted(
+      final List<T> items, final Function<T, U> propertyExtractor) {
+    assertThatIsSortedBy(items, propertyExtractor, Comparator.reverseOrder());
+  }
+
+  static <T, U extends Comparable<U>> void assertThatIsSortedBy(
+      final List<T> items, final Function<T, U> propertyExtractor, final Comparator<U> comparator) {
+    final var sorted =
+        items.stream().map(propertyExtractor).filter(Objects::nonNull).sorted(comparator).toList();
+    assertThat(items.stream().map(propertyExtractor).filter(Objects::nonNull).toList())
+        .containsExactlyElementsOf(sorted);
+  }
+}
